Move mosaic functions into their specific image types
